+++ telegram/index.js
@@ -1,0 +1,3 @@
+const startTelegramBot = require('./start_telegram_bot');
+
+module.exports = {startTelegramBot};